I kidna forgot what i just just did, but I am using vim to write this commit.
It is actually kinda cool, and I should start to try and use VIM more.  The real problem is that I cannot use any mouse commands to alter code.  I have to physically go back using the cursor, which is a huge pain in the ass.
I guess that is why they made selection and navigation short cuts.